+++ CHANGELOG.md
@@ -8,6 +8,7 @@ in 0.1 minor versions.
 To get the diff for a specific change, go to https://github.com/superdesk/web-publisher/commit/XXX where XXX is the change hash
 
 * 0.2.0
+ * feature [#218] Assign article to route of type content automatically when article is published
  * feature [#215] Filter articles by metadata in gimmelist
  * feature [#213] Implement and expose article's keywords
  * feature [#211] Create Automatic list widget